feat(search): promote session-rollup search and port saved searches
Retires the legacy trace-flat /search in favor of the session-rollup
view that was running behind the session-search-v2 flag. Saved searches
and the Export Traces / Add to Dataset bulk actions are ported over so
the new /search is a strict superset of what the old route offered. The
Live Taxonomy panel is dropped (and its supporting feature flag and
web-side hooks removed) since clusters are trace-centric and don't map
cleanly onto sessions.

Selection in the session view is already trace-id-based via
useSessionSelectionAdapter, so existing bulk-export plumbing works
unchanged; useTracesCount feeds the real trace count for select-all.


WIP

fix(search): apply hasLlmActivity default to bulk-action filters

The sessions hooks (useSessionsInfiniteScroll, useSessionsCount) apply
`withSessionDefaults` internally, but the trace-side surfaces on the
search page were sending the raw filter set:

- useTracesCount produced an inflated trace count, so the "Export
  Traces (N)" chip exceeded what was actually visible in the list.
- enqueueTracesExport re-ran server-side without the hasLlmActivity
  clause for Select-All exports, sweeping orphan-fragment traces the
  user never saw.
- AddToDatasetModal had the same hole.

Hoists a single `effectiveFilters = withSessionDefaults(filters)` memo
and routes the three callsites through it. `withSessionDefaults` is
now exported from `sessions.collection.ts`.

Also broadens `NUMBER_RANGE_FIELDS` percentile attachment to accept
both trace and session percentile fields — today the lists are
identical, but the trace-only guard would silently drop a future
session-only percentile field's UI.


fix(traces): apply hasLlmActivity default on Sessions tab bulk actions

Same orphan-fragment leak `/search` had, now patched on the project
root's Sessions tab. `useTracesCount`, `enqueueTracesExport`, and
`AddToDatasetModal` were sending the raw filter set; on the Sessions
tab they need `withSessionDefaults(filters)` so a Select-All export
matches what the user actually sees in the list. The conditional is
gated on `activeTab === "sessions"` since `hasLlmActivity` is a
session-only synthetic and isn't part of the trace filter registry.


Show only live session status

fix(search): open orphan search hits as their trace instead of "not found"

The session search runs over the trace search index. Two cases produce
search hits whose `sessionId` doesn't resolve to a row in the `sessions`
table — pre-migration traces written before migration 00016 #3224 (when
sessions_mv filtered `WHERE session_id != ''`), and the structural race
between trace ingestion and sessions_mv replication. In both, the list
synthesizes a placeholder Session via `toOrphanSession(row)` whose
sessionId equals `toString(trace_id)`.

Clicking such a row currently lands on "Session not found." because
`findBySessionId` against the sessions table finds nothing. Falls back
to opening the trace directly using the same id — orphans always have
a resolvable trace under that id. The "View session" back affordance is
suppressed when the session is missing (nothing to back to), and the
final TraceDetailBody renders its own not-found UI if the trace also
doesn't exist.
