--- apps/web/src/domains/taxonomy/taxonomy.collection.ts
@@ -1,13 +1,0 @@
-import { useQuery } from "@tanstack/react-query"
-import { getTaxonomyOverview } from "./taxonomy.functions.ts"
-
-const taxonomyOverviewQueryKey = (projectId: string) => ["taxonomyOverview", projectId] as const
-
-export function useTaxonomyOverview(projectId: string) {
-  return useQuery({
-    queryKey: taxonomyOverviewQueryKey(projectId),
-    queryFn: () => getTaxonomyOverview({ data: { projectId } }),
-    staleTime: 30_000,
-    enabled: projectId.length > 0,
-  })
-}
