+++ apps/web/src/routes/_authenticated/projects/$projectSlug/-components/save-search-modal.tsx
@@ -4,10 +4,10 @@ import { useForm } from "@tanstack/react-form"
 import {
   useCreateSavedSearch,
   useUpdateSavedSearch,
-} from "../../../../../../domains/saved-searches/saved-searches.collection.ts"
-import type { SavedSearchRecord } from "../../../../../../domains/saved-searches/saved-searches.functions.ts"
-import { toUserMessage } from "../../../../../../lib/errors.ts"
-import { createFormSubmitHandler, fieldErrorsAsStrings } from "../../../../../../lib/form-server-action.ts"
+} from "../../../../../domains/saved-searches/saved-searches.collection.ts"
+import type { SavedSearchRecord } from "../../../../../domains/saved-searches/saved-searches.functions.ts"
+import { toUserMessage } from "../../../../../lib/errors.ts"
+import { createFormSubmitHandler, fieldErrorsAsStrings } from "../../../../../lib/form-server-action.ts"
 
 interface BaseProps {
   readonly open: boolean
