+++ apps/web/src/routes/_authenticated/projects/$projectSlug/-components/saved-searches-list.tsx
@@ -13,18 +13,18 @@ import {
 import { useRouter } from "@tanstack/react-router"
 import { FilterIcon, SearchIcon, SparklesIcon } from "lucide-react"
 import { useState } from "react"
-import { MemberSelector } from "../../../../../../components/member-selector.tsx"
+import { MemberSelector } from "../../../../../components/member-selector.tsx"
 import {
   type SavedSearchAggregates,
   useDeleteSavedSearch,
   useSavedSearchAggregates,
   useSavedSearchesList,
   useUpdateSavedSearch,
-} from "../../../../../../domains/saved-searches/saved-searches.collection.ts"
-import type { SavedSearchRecord } from "../../../../../../domains/saved-searches/saved-searches.functions.ts"
-import { toUserMessage } from "../../../../../../lib/errors.ts"
-import { serializeFilters } from "../../-components/trace-page-state.ts"
+} from "../../../../../domains/saved-searches/saved-searches.collection.ts"
+import type { SavedSearchRecord } from "../../../../../domains/saved-searches/saved-searches.functions.ts"
+import { toUserMessage } from "../../../../../lib/errors.ts"
 import { SaveSearchModal } from "./save-search-modal.tsx"
+import { serializeFilters } from "./trace-page-state.ts"
 
 const dateFormatter = new Intl.DateTimeFormat(undefined, { year: "numeric", month: "short", day: "numeric" })
 const numberFormatter = new Intl.NumberFormat()
