+++ apps/web/src/routes/_authenticated/projects/$projectSlug/-components/search-syntax-legend.tsx
