+++ apps/web/src/routes/_authenticated/projects/$projectSlug/-components/session-detail-drawer/session-status-pill.tsx
@@ -1,7 +1,11 @@
-import { Status } from "@repo/ui"
+import { Status, Tooltip } from "@repo/ui"
 import type { SessionStatus } from "../../../../../../domains/sessions/sessions.collection.ts"
 
-/** Green "Live · 3m ago" / muted "Idle · 42m ago" pill from the session status. */
+/**
+ * Green "Live · 3m ago" pill, shown only while the session is live. Idle
+ * sessions render nothing — the surrounding row already conveys the last
+ * activity timestamp.
+ */
 export function SessionStatusPill({
   status,
   lastActivity,
@@ -9,10 +13,11 @@ export function SessionStatusPill({
   readonly status: SessionStatus
   readonly lastActivity: string
 }) {
+  if (status !== "live") return null
+
   return (
-    <Status
-      variant={status === "live" ? "success" : "neutral"}
-      label={`${status === "live" ? "Live" : "Idle"} · ${lastActivity}`}
-    />
+    <Tooltip asChild trigger={<Status variant="success" label={`Live · ${lastActivity}`} />}>
+      This session had activity in the last 5 minutes — more may still be coming in.
+    </Tooltip>
   )
 }
