+++ tools/live-seeds/src/scripts/send-session-search-qa.ts
@@ -142,9 +142,7 @@ void sendLiveSeedData(options)
     console.log("✓ Done. Next steps:")
     console.log("  1. Wait ~10 seconds for the trace-search worker to index the new traces")
     console.log("     (debounce after trace-end:run — see apps/workers/src/workers/trace-search.ts).")
-    console.log("  2. Make sure the session-search-v2 feature flag is enabled for your org via")
-    console.log("     /backoffice/feature-flags/.")
-    console.log("  3. Open /projects/<project-slug>/session-search/ and run the queries listed in")
+    console.log("  2. Open /projects/<project-slug>/session-search/ and run the queries listed in")
     console.log("     dev-docs/qa/session-search.md.")
   })
   .catch((error: unknown) => {
